requirements.txt
streamlit==1.38.0
transformers==4.44.2
accelerate>=0.33.0
sentencepiece
torch  # On Windows CPU, pip will fetch CPU build
langdetect==1.0.9
huggingface-hub>=0.24.6
python-dotenv==1.0.1
pydantic>=2.8.2
requests>=2.32.3
